Remove .closeOnJvmShutdown() mapdb calls when using database files as this causes a race condition on JVM shutdown between mapdb trying to close the database files and QORA. Now responsibility for safely closing the files lies solely with QORA.
Some work been done to the networking code as it seems to fight with itself.
e.g. when pinging a peer fails, the Peer class tries to stop the pinger
but the Pinger class causes an exception

Also trying to improve the overall shutdown sequence by stopping
the BlockGenerator class as it uses the (supposedly stopped) networking stack
and tries to make mapdb calls - possibly after the database is closed.

Seems safe to close QORA using GUI methods (X close button, "Exit", etc.)
but not non-GUI methods like "kill" or control-C.