Unify name storage data extraction from JSON
This is all in utils/StorageUtils.java:

The previous code for extracting data for name storage from transaction JSON blindly assumed the data would be a string-encapsulated JSONObject.
Also this code was duplicated in various places.

Now this has been refactored into one place: getDataByKey()
This function is more permissive in that it allows optional string encapsulation but also more restrictive in that it will return null if the data is not suitable.

This should help with issues #57 and #74.

Other functions in the same file have been tidied up.

+ version bump as this has been blocking many users from using client (unless they wipe database)